requirements.txt
asgiref==3.6.0
attrs==22.2.0
Babel==2.11.0
black==22.12.0
click==8.1.3
Django==4.1.5
django-money==3.0.0
djangorestframework==3.14.0
exceptiongroup==1.1.0
iniconfig==2.0.0
mypy-extensions==0.4.3
packaging==23.0
pathspec==0.11.0
platformdirs==2.6.2
pluggy==1.0.0
py-moneyed==2.0
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
